actualizado

+++ templates_c/14f3d9d14043f78408eee90cb07417814a8ebaba_0.file.albumsList.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.2.1, created on 2022-10-15 02:10:03
+/* Smarty version 4.2.1, created on 2022-10-17 17:17:09
   from 'C:\xampp\htdocs\TPEweb2\templates\albumsList.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_6349fa5b3cc9b1_09894985',
+  'unifunc' => 'content_634d71f5dd9760_06914314',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     '14f3d9d14043f78408eee90cb07417814a8ebaba' => 
     array (
       0 => 'C:\\xampp\\htdocs\\TPEweb2\\templates\\albumsList.tpl',
-      1 => 1665764307,
+      1 => 1666019663,
       2 => 'file',
     ),
   ),
@@ -22,7 +22,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:templates/footer.tpl' => 1,
   ),
 ),false)) {
-function content_6349fa5b3cc9b1_09894985 (Smarty_Internal_Template $_smarty_tpl) {
+function content_634d71f5dd9760_06914314 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_subTemplateRender("file:templates/header.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
 
@@ -43,30 +43,40 @@ $_smarty_tpl->tpl_vars['album']->do_else = false;
 " class="w-100">
                         </a>
                     </div>
-                    <div class="card-body p-2">
+                    <div class="card-body p-2 ">
                         <h2 class="h5 card-text text-center text-muted"><?php echo $_smarty_tpl->tpl_vars['album']->value->nombre;?>
-<a href="deleteAlbum/<?php echo $_smarty_tpl->tpl_vars['album']->value->id;?>
-" ><button type="button" class="btn btn-danger" >X</button></a></h2>
+
+
+                        </h2>
+                        <?php if ((isset($_SESSION['USER_ID']))) {?>
+                            <div class="justify-content-around d-flex">
+                                <a href="adminEdit"><button type="button" class="btn text-warning btn-light">Edit</button></a>
+                                <a href="deleteAlbum/<?php echo $_smarty_tpl->tpl_vars['album']->value->id;?>
+"><button type="button" class="btn text-danger btn-light">Delete</button></a>
+                            </div>
+                        <?php }?>
                     </div>
                 </div>
             </div>
-            
+
         <?php
 }
 $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
-        <div class="col">
-            <div class="card p-1 bg-light mx-2">
-                <div>
-                    <a href="admin">
-                        <img src="https://us.123rf.com/450wm/siamimages/siamimages1611/siamimages161100055/65441642-a%C3%B1adir-signo-m%C3%A1s-ic%C3%B3nico-dise%C3%B1o.jpg?ver=6"
-                            class="images img-thumbnail" height="230">
-                    </a>
-                </div>
-                <div class="card-body p-2">
-                    <h2 class="h5 card-text text-center text-muted">Agregar album</h2>
+        <?php if ((isset($_SESSION['USER_ID']))) {?>
+            <div class="col">
+                <div class="card p-1 bg-light mx-2">
+                    <div>
+                        <a href="adminAdd">
+                            <img src="https://us.123rf.com/450wm/siamimages/siamimages1611/siamimages161100055/65441642-a%C3%B1adir-signo-m%C3%A1s-ic%C3%B3nico-dise%C3%B1o.jpg?ver=6"
+                                class="images img-thumbnail" height="230">
+                        </a>
+                    </div>
+                    <div class="card-body p-2 ">
+                        <h2 class="h5 card-text text-center text-muted">Agregar album</h2>
+                    </div>
                 </div>
             </div>
-        </div>
+        <?php }?>
     </div>
 </div>
